Storage: warnings - careful with mixing 'unsigned int' with 'size_t' for offset/sizes, it invites a lot of warnings from 64-bit MSC, which can only be solved by ugly casts.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@62732 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Storage/testcase/VDScriptAst.cpp
@@ -346,7 +346,7 @@ DECLHIDDEN(PVDSCRIPTASTCORE) vdScriptAstNodeAlloc(VDSCRIPTASTCLASS enmClass)
     return pAstNode;
 }
 
-DECLHIDDEN(PVDSCRIPTASTIDE) vdScriptAstNodeIdeAlloc(unsigned cchIde)
+DECLHIDDEN(PVDSCRIPTASTIDE) vdScriptAstNodeIdeAlloc(size_t cchIde)
 {
     PVDSCRIPTASTIDE pAstNode = (PVDSCRIPTASTIDE)RTMemAllocZ(RT_OFFSETOF(VDSCRIPTASTIDE, aszIde[cchIde + 1]));
     if (pAstNode)
